requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
pillow
fastparquet
notebook
jupyterlab
split-folders